Fixed wrong prototype on a method

+++ include/nexrad/color.h
@@ -25,7 +25,7 @@ void nexrad_color_table_store_entry(nexrad_color_table *table,
 
 nexrad_color_table *nexrad_color_table_open(const char *path);
 
-nexrad_color_table_entry *nexrad_color_table_get_entries(nexrad_color_table *table, size_t *size);
+nexrad_color_table_entry *nexrad_color_table_get_entries(nexrad_color_table *table, uint8_t *size);
 
 int nexrad_color_table_save(nexrad_color_table *table, const char *path);
